Refactor code and using Liquibase maven

+++ source/user-base-auth/src/main/java/com/base/auth/controller/CategoryController.java
@@ -3,18 +3,21 @@ package com.base.auth.controller;
 import com.base.auth.dto.ApiResponse;
 import com.base.auth.dto.ErrorCode;
 import com.base.auth.dto.ResponseListDto;
+import com.base.auth.dto.category.CategoryDto;
 import com.base.auth.form.category.CreateCategoryForm;
 import com.base.auth.form.category.UpdateCategoryForm;
 import com.base.auth.mapper.CategoryMapper;
 import com.base.auth.model.Category;
 import com.base.auth.model.criteria.CategoryCriteria;
 import com.base.auth.repository.CategoryRepository;
+import com.base.auth.repository.NewsRepository;
 import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
@@ -31,9 +34,13 @@ public class CategoryController extends ABasicController {
     private CategoryRepository categoryRepository;
 
     @Autowired
+    private NewsRepository newsRepository;
+
+    @Autowired
     private CategoryMapper categoryMapper;
 
     @PostMapping(value = "/create", produces = MediaType.APPLICATION_JSON_VALUE)
+    @PreAuthorize("hasRole('CATE_C')")
     public ApiResponse<String> createCategory(@Valid @RequestBody CreateCategoryForm createCategoryForm, BindingResult bindingResult) {
         ApiResponse<String> apiMessageDto = new ApiResponse<>();
         Category category = categoryRepository.findCategoryByName(createCategoryForm.getName());
@@ -50,6 +57,7 @@ public class CategoryController extends ABasicController {
     }
 
     @PutMapping(value = "/update", produces = MediaType.APPLICATION_JSON_VALUE)
+    @PreAuthorize("hasRole('CATE_U')")
     public ApiResponse<String> updateCategory(@Valid @RequestBody UpdateCategoryForm updateCategoryForm, BindingResult bindingResult) {
         ApiResponse<String> apiMessageDto = new ApiResponse<>();
         Category category = categoryRepository.findById(updateCategoryForm.getCategoryId()).orElse(null);
@@ -66,31 +74,42 @@ public class CategoryController extends ABasicController {
     }
 
     @GetMapping(value = "/get/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ApiResponse<Category> get(@PathVariable("id") Long id) {
+    @PreAuthorize("hasRole('CATE_V')")
+    public ApiResponse<CategoryDto> get(@PathVariable("id") Long id) {
+        ApiResponse<CategoryDto> apiMessageDto = new ApiResponse<>();
+
         Category category = categoryRepository.findById(id).orElse(null);
-        ApiResponse<Category> apiMessageDto = new ApiResponse<>();
-        apiMessageDto.setData(category);
+        if (category == null) {
+            apiMessageDto.setResult(false);
+            apiMessageDto.setCode(ErrorCode.CATEGORY_ERROR_NOT_FOUND);
+            return apiMessageDto;
+        }
+
+        CategoryDto categoryDto = categoryMapper.fromEntityToCategoryDto(category);
+        apiMessageDto.setData(categoryDto);
         apiMessageDto.setMessage("Get category success");
         return apiMessageDto;
     }
 
     @GetMapping(value = "/list", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ApiResponse<ResponseListDto<List<Category>>> get(CategoryCriteria criteria, Pageable pageable) {
+    @PreAuthorize("hasRole('CATE_L')")
+    public ApiResponse<ResponseListDto<List<CategoryDto>>> get(CategoryCriteria criteria, Pageable pageable) {
         Page<Category> pageData = categoryRepository.findAll(criteria.getSpecification(), pageable);
 
-        ResponseListDto<List<Category>> responseListDto = new ResponseListDto<>();
-        responseListDto.setContent(pageData.getContent());
+        ResponseListDto<List<CategoryDto>> responseListDto = new ResponseListDto<>();
+        responseListDto.setContent(categoryMapper.fromEntityToDtoList(pageData.getContent()));
         responseListDto.setTotalElements(pageData.getTotalElements());
         responseListDto.setTotalPages(pageData.getTotalPages());
 
-        ApiResponse<ResponseListDto<List<Category>>> apiMessageDto = new ApiResponse<>();
+        ApiResponse<ResponseListDto<List<CategoryDto>>> apiMessageDto = new ApiResponse<>();
         apiMessageDto.setData(responseListDto);
         apiMessageDto.setMessage("Get list categories success");
         return apiMessageDto;
     }
 
     @DeleteMapping(value = "/delete/{id}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ApiResponse<String> updateCategory(@PathVariable("id") Long id) {
+    @PreAuthorize("hasRole('CATE_D')")
+    public ApiResponse<String> delete(@PathVariable("id") Long id) {
         ApiResponse<String> apiMessageDto = new ApiResponse<>();
         Category category = categoryRepository.findById(id).orElse(null);
         if (category == null) {
@@ -98,6 +117,13 @@ public class CategoryController extends ABasicController {
             apiMessageDto.setCode(ErrorCode.CATEGORY_ERROR_NOT_FOUND);
             return apiMessageDto;
         }
+
+        if (newsRepository.countNewsByCategoryId(category.getId()) > 0) {
+            apiMessageDto.setResult(false);
+            apiMessageDto.setCode(ErrorCode.CATEGORY_ERROR_CANT_DELETE_RELATIONSHIP_WITH_NEWS);
+            return apiMessageDto;
+        }
+
         categoryRepository.deleteById(id);
 
         apiMessageDto.setMessage("Delete category success");
